:recycle: Refactor utils, let user deposit on /send when not exists
