+++ backend/src/client.js
@@ -23,6 +23,7 @@ client.once('ready', async () => {
 
 client.on('interactionCreate', async (interaction) => {
   const command = client.commands.get(interaction.commandName);
+  console.log(interaction.commandName);
 
   if (!command) return;
