+++ backend/src/utils/api.js
@@ -1,0 +1,9 @@
+import axios from 'axios';
+import qs from 'qs';
+
+import { ENDPOINT } from '../config.js';
+
+export default axios.create({
+  baseURL: ENDPOINT,
+  paramsSerializer: (params) => qs.stringify(params, { arrayFormat: 'repeat' }),
+});
